requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
pydantic==2.10.6
psycopg[binary]==3.2.6
python-dotenv==1.0.1
